Crud de doctor, mejorable, crud de paciente mejorable
