+++ root/config/app.php
@@ -1,8 +1,13 @@
 <?php
-define('ROOT_PATH', 'C:/xampp/htdocs/HealthConnection/root/');
-define('DB_CONNECTION_PATH', ROOT_PATH . 'config/databaseConexion.php');
-define('PUBLIC_PATH', ROOT_PATH . 'public/');
-define('SRC_PATH', ROOT_PATH . 'src/');
-define('CONTROLLERS_PATH', SRC_PATH . 'controllers/');
-define('MODELS_PATH', SRC_PATH . 'models/');
-define('VIEWS_PATH', SRC_PATH . 'views/');
+
+$ROOT_PATH = 'C:/xampp/htdocs/HealthConnection/root/';
+$CONFIG_PATH = $ROOT_PATH . 'config/';
+$CONEXION = $CONFIG_PATH . 'databaseConexion.php';
+$PUBLIC = $ROOT_PATH . 'public/';
+$SRC = $ROOT_PATH . 'src/';
+$MODELS = $SRC . 'models/';
+$CONTROLLERS = $SRC . 'controllers/';
+$VIEWS_PATH = $SRC . 'views/';
+
+
+
