+++ root/public/css/admin/appointment/formularioCita.css
@@ -1,0 +1,99 @@
+@import url('https://fonts.googleapis.com/css2?family=Coustard:wght@400;900&family=Nunito:ital,wght@0,200..1000;1,200..1000&family=Quicksand:wght@300..700&family=Roboto+Mono:ital,wght@0,100..700;1,100..700&family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap');
+
+        :root{
+            --color-principal: #ff914d;
+            --color-secundario:#6C5F5B;
+            --color-complementario:#4F4A45;
+            --negro:#151515;
+            --gris-clarito:#9f9e9ed4;
+            --white:#FFFFFF;
+            --fuente-principal:"Roboto Mono", monospace;
+            --fuente-secundaria:'Nunito',sans-serif;
+            --small-font:14px;
+            --medium-font:16px;
+            --large-font:18px;
+            --gris-fondo:#e8e5e5;
+            --logo-font:"Coustard", serif;
+        }
+        
+        body {
+            background-color: var(--gris-fondo);
+            margin: 0;
+            padding: 0;
+            font-family: var(--fuente-principal);
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            min-height: 100vh;
+            flex-direction: column;
+            text-align: center;
+        }
+        
+        .appointment-container {
+            background-color: var(--white);
+            border: 1px solid var(--color-secundario);
+            border-radius: 10px;
+            padding: 20px;
+            box-shadow: 10px 8px 16px var(--color-complementario);
+        }
+
+        .form {
+            background-color: var(--color-principal);
+            padding: 10px;
+            border-radius: 10px;
+            text-align: center;
+        }
+
+        .form h2 {
+            color: var(--white);
+            font-family: var(--fuente-principal);
+        }
+
+        .appointment-form {
+            padding: 20px;
+            border-radius: 5px;
+            background-color: var(--white);
+        }
+
+        .form-group select ,
+        .form-group input[type="date"],
+        .form-group input[type="time"],
+        .form-group input[type="text"] {
+            width: 60%;
+            padding: 10px;
+            border: 2px solid var(--color-principal);
+            border-radius: 5px;
+            font-family: var(--fuente-principal);
+            font-weight: 600;
+        }
+
+        button[type="submit"] {
+            background-color: var(--color-principal);
+            color: white;
+            border: none;
+            padding: 10px 20px;
+            border-radius: 5px;
+            cursor: pointer;
+            width: 100%;
+            font-family: var(--fuente-principal);
+            font-size: var(--large-font);
+            font-weight: bold;
+        }
+
+        button[type="submit"]:hover {
+            background-color: var(--white);
+            color: var(--color-principal);
+            border: 1px solid var(--color-principal);
+        }
+
+        .bg-body-secondary {
+            margin-top: 100px;
+        }
+
+        .footer {
+            background-color: #333;
+            color: white;
+            text-align: center;
+            padding: 20px;
+            width: 100%;
+        }
