+++ root/public/css/admin/verRegistros.css
@@ -26,6 +26,7 @@ body {
     margin-top: 100px;
 }
 
+
 .container-record {
     padding: 20px;
     margin: 20px auto;
