+++ root/public/css/admin/verSolicitudes.css
@@ -1,0 +1,127 @@
+:root {
+    --color-principal: #ff914d;
+    --color-secundario: #6C5F5B;
+    --color-complementario: #4F4A45;
+    --negro: #151515;
+    --gris-clarito: #9f9e9ed4;
+    --white: #FFFFFF;
+    --fuente-principal: "Roboto Mono", monospace;
+    --fuente-secundaria: 'Nunito', sans-serif;
+    --small-font: 14px;
+    --medium-font: 16px;
+    --large-font: 18px;
+    --gris-fondo: #e8e5e5;
+}
+
+body {
+    font-family: var(--fuente-principal);            
+    margin: 0;
+    padding: 0;
+}
+
+.bg-body-secondary {
+    margin-top: 100px;
+}  
+
+.container-record {
+    width: 90%;
+    max-width: 90%;
+    padding: 20px;
+    margin: 20px auto;
+    background-color: var(--white);
+    border: 1px solid var(--color-secundario);
+    border-radius: 8px;
+    box-shadow: 10px 8px 16px var(--color-complementario);
+    flex-grow: 1;
+}
+
+.nav-menu {
+    display: flex;
+    justify-content: space-between;
+    flex-wrap: wrap;
+    margin-bottom: 20px;
+    gap: 10px;
+}
+
+.nav-menu button {
+    flex: 1;
+    background-color: var(--white);
+    border: 1px solid #ccc;
+    padding: 10px 20px;
+    border-radius: 5px;
+    cursor: pointer;
+    transition: background-color 0.3s, color 0.3s, border 0.3s;
+}
+
+.nav-menu button:hover {
+    background-color: var(--color-principal);
+    color: var(--white);
+    border: 2px solid var(--color-principal);
+}
+
+.content-section {
+    display: none;
+    width: 100%;
+}
+
+.content-section.active {
+    display: block;
+}
+
+.title {
+    text-align: center;
+    background-color: var(--color-principal);
+    padding: 10px;
+    border-radius: 8px;
+    color: var(--white);
+    font-size: var(--large-font);
+    margin-bottom: 20px;
+}
+
+.primary-button {
+    background-color: var(--white);
+    color: var(--color-principal);
+    border: 2px solid var(--color-principal);
+    font-size: var(--large-font);
+    border-radius: 8px;
+    cursor: pointer;
+    transition: background-color 0.3s, color 0.3s;
+}
+
+.primary-button:hover {
+    background-color: var(--color-principal);
+    color: var(--white);
+}
+
+.appointment-card, .payment-card {
+    border: 2px solid var(--color-principal);
+    padding: 20px;
+    background-color: var(--white);
+    border-radius: 8px;
+    box-shadow: 2px 2px 5px rgba(0, 0, 0, 0.1);
+    margin-bottom: 20px;
+    word-wrap: break-word;
+}
+
+.actions {
+    display: flex;
+    justify-content: space-around;
+}
+
+.search-bar {
+    margin: 20px 0;
+}
+
+/* Hacer que la tabla sea responsive */
+.table-responsive {
+    overflow-x: auto;
+}
+
+@media (max-width: 768px) {
+    .nav-menu {
+        flex-direction: column;
+    }
+    .container-record {
+        max-width: 100%;
+    }
+}
