--- root/src/controllers/consultarRegistros.php
@@ -1,4 +1,0 @@
-<?php
-include_once "../../../config/databaseConexion.php";
-$sentencia= $database->query("SELECT * FROM Paciente;");
-$paciente= $sentencia->fetchAll(PDO::FETCH_OBJ);
