+++ root/src/views/admin/verSolicitudes.php
@@ -1,5 +1,5 @@
 <?php 
-include_once '../../controllers/consultarCitas.php'; 
+include_once '../../controllers/appointmentController/consultarCitas.php'; 
 $baseUrlPublic = '/HealthConnection/root/public/';
 $baseUrlSrc = 'C:/xampp/htdocs/HealthConnection/root/src/';
 $baseUrlSrcFooter = 'C:/xampp/htdocs/HealthConnection/root/src/';
@@ -17,142 +17,14 @@ $baseUrlSrcFooter = 'C:/xampp/htdocs/HealthConnection/root/src/';
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/css/bootstrap.min.css">
     <!-- Bootstrap JS -->
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js"></script>
+    <link rel="stylesheet" href="../../../public/css/admin/verSolicitudes.css">
     
-    <style>
-        :root {
-            --color-principal: #ff914d;
-            --color-secundario: #6C5F5B;
-            --color-complementario: #4F4A45;
-            --negro: #151515;
-            --gris-clarito: #9f9e9ed4;
-            --white: #FFFFFF;
-            --fuente-principal: "Roboto Mono", monospace;
-            --fuente-secundaria: 'Nunito', sans-serif;
-            --small-font: 14px;
-            --medium-font: 16px;
-            --large-font: 18px;
-            --gris-fondo: #e8e5e5;
-        }
-
-        body {
-            font-family: var(--fuente-principal);            
-            margin: 0;
-            padding: 0;
-        }
-
-        .bg-body-secondary {
-            margin-top: 100px;
-        }  
-
-        .container-record {
-            width: 90%;
-            max-width: 90%;
-            padding: 20px;
-            margin: 20px auto;
-            background-color: var(--white);
-            border: 1px solid var(--color-secundario);
-            border-radius: 8px;
-            box-shadow: 10px 8px 16px var(--color-complementario);
-            flex-grow: 1;
-        }
-
-        .nav-menu {
-            display: flex;
-            justify-content: space-between;
-            flex-wrap: wrap;
-            margin-bottom: 20px;
-            gap: 10px;
-        }
-
-        .nav-menu button {
-            flex: 1;
-            background-color: var(--white);
-            border: 1px solid #ccc;
-            padding: 10px 20px;
-            border-radius: 5px;
-            cursor: pointer;
-            transition: background-color 0.3s, color 0.3s, border 0.3s;
-        }
-
-        .nav-menu button:hover {
-            background-color: var(--color-principal);
-            color: var(--white);
-            border: 2px solid var(--color-principal);
-        }
-
-        .content-section {
-            display: none;
-            width: 100%;
-        }
-
-        .content-section.active {
-            display: block;
-        }
-
-        .title {
-            text-align: center;
-            background-color: var(--color-principal);
-            padding: 10px;
-            border-radius: 8px;
-            color: var(--white);
-            font-size: var(--large-font);
-            margin-bottom: 20px;
-        }
-
-        .primary-button {
-            background-color: var(--white);
-            color: var(--color-principal);
-            border: 2px solid var(--color-principal);
-            font-size: var(--large-font);
-            border-radius: 8px;
-            cursor: pointer;
-            transition: background-color 0.3s, color 0.3s;
-        }
-
-        .primary-button:hover {
-            background-color: var(--color-principal);
-            color: var(--white);
-        }
-
-        .appointment-card, .payment-card {
-            border: 2px solid var(--color-principal);
-            padding: 20px;
-            background-color: var(--white);
-            border-radius: 8px;
-            box-shadow: 2px 2px 5px rgba(0, 0, 0, 0.1);
-            margin-bottom: 20px;
-            word-wrap: break-word;
-        }
-
-        .actions {
-            display: flex;
-            justify-content: space-around;
-        }
-
-        .search-bar {
-            margin: 20px 0;
-        }
-
-        /* Hacer que la tabla sea responsive */
-        .table-responsive {
-            overflow-x: auto;
-        }
-
-        @media (max-width: 768px) {
-            .nav-menu {
-                flex-direction: column;
-            }
-            .container-record {
-                max-width: 100%;
-            }
-        }
-    </style>
 </head>
 
 <body class="bg-body-secondary">
 
     <!-- Navbar -->
-    <?php include $baseUrlSrc . 'views/layouts/navbar.php'; ?>
+    <?php include  '../../views/layouts/navbar.php'; ?>
 
     <main class="container-record">
         <!-- Menú de navegación -->
@@ -214,7 +86,7 @@ $baseUrlSrcFooter = 'C:/xampp/htdocs/HealthConnection/root/src/';
                                             </div>
                                         </div>
 
-                                        <a class="btn btn-primary" href="editarCita.php?id_cita=<?php echo htmlspecialchars($cita->id_cita); ?>">Editar</a>
+                                        <a class="btn btn-primary" href="./appointment/editarCita.php?id_cita=<?php echo htmlspecialchars($cita->id_cita); ?>">Editar</a>
                                         <a type="button" class="btn btn-danger" data-bs-toggle="modal" data-bs-target="#confirmModal<?php echo htmlspecialchars($cita->id_cita); ?>">Eliminar</a>
 
                                         <!-- Modal Confirmación Eliminación -->
@@ -231,7 +103,7 @@ $baseUrlSrcFooter = 'C:/xampp/htdocs/HealthConnection/root/src/';
                                                         ¿Estás seguro de que deseas eliminar esta cita?
                                                     </div>
                                                     <div class="modal-footer">
-                                                        <form action="../../controllers/deleteCita.php" method="POST">
+                                                        <form action="../../controllers/appointmentController/deleteCita.php" method="POST">
                                                             <input type="hidden" name="id_cita" value="<?php echo htmlspecialchars($cita->id_cita); ?>">
                                                             <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Cancelar</button>
                                                             <button type="submit" class="btn btn-danger">Eliminar</button>
@@ -277,7 +149,7 @@ $baseUrlSrcFooter = 'C:/xampp/htdocs/HealthConnection/root/src/';
         </main>
 
         <!-- Footer -->
-        <?php include $baseUrlSrcFooter . 'views/layouts/footer.php'; ?>
+        <?php include  '../../views/layouts/footer.php'; ?>
 
     <script>
         function showSection(sectionId) {
